+++ src/Azure.DevOps.PSModule/Public/Graph/Memberships/Get-AdoMembership.ps1
@@ -1,10 +1,11 @@
 ﻿function Get-AdoMembership {
     <#
     .SYNOPSIS
-        Get the membership relationship between a subject and a container in Azure DevOps.
+        Get membership relationships
 
     .DESCRIPTION
-        This cmdlet retrieves the membership relationship between a specified subject and container in Azure DevOps.
+        This cmdlet retrieves the membership relationships between a specified subject and container in Azure DevOps or
+        get all the memberships where this descriptor is a member in the relationship.
 
     .PARAMETER CollectionUri
         Optional. The collection URI of the Azure DevOps collection/organization, e.g., https://vssps.dev.azure.com/my-org.
@@ -13,16 +14,26 @@
         Mandatory. A descriptor to the child subject in the relationship.
 
     .PARAMETER ContainerDescriptor
-        Mandatory. A descriptor to the container in the relationship.
+        Optional. A descriptor to the container in the relationship.
+
+    .PARAMETER Depth
+        Optional. The depth of memberships to retrieve when ContainerDescriptor is not specified. Default is 1.
+
+    .PARAMETER Direction
+        Optional. The direction of memberships to retrieve when ContainerDescriptor is not specified.
+
+        The default value for direction is 'up' meaning return all memberships where the subject is a member (e.g. all groups the subject is a member of).
+        Alternatively, passing the direction as 'down' will return all memberships where the subject is a container (e.g. all members of the subject group).
 
     .PARAMETER Version
-        Optional. The API version to use for the request. Default is '7.1'.
+        Optional. The API version to use for the request. Default is '7.1-preview.1'.
 
     .OUTPUTS
         PSCustomObject
 
     .LINK
-        https://learn.microsoft.com/en-us/rest/api/azure/devops/graph/memberships/get
+        - https://learn.microsoft.com/en-us/rest/api/azure/devops/graph/memberships/get
+        - https://learn.microsoft.com/en-us/rest/api/azure/devops/graph/memberships/list
 
     .EXAMPLE
         $params = @{
@@ -42,6 +53,28 @@
         @('aadgp.00000000-0000-0000-0000-000000000002', 'aadgp.00000000-0000-0000-0000-000000000003') | Get-AdoMembership @params
 
         Retrieves the membership relationships for multiple subjects demonstrating pipeline input.
+
+    .EXAMPLE
+        $params = @{
+            CollectionUri     = 'https://vssps.dev.azure.com/my-org'
+            SubjectDescriptor = 'aadgp.00000000-0000-0000-0000-000000000000'
+            Depth             = 2
+            Direction         = 'up'
+        }
+        Get-AdoMembership @params
+
+        Retrieves all groups for a user with a depth of 2.
+
+    .EXAMPLE
+        $params = @{
+            CollectionUri     = 'https://vssps.dev.azure.com/my-org'
+            SubjectDescriptor = 'aadgp.00000000-0000-0000-0000-000000000000'
+            Depth             = 2
+            Direction         = 'down'
+        }
+        Get-AdoMembership @params
+
+        Retrieves all memberships of a group with a depth of 2.
     #>
     [CmdletBinding()]
     param (
@@ -52,8 +85,15 @@
         [Parameter(Mandatory, ValueFromPipelineByPropertyName, ValueFromPipeline)]
         [string[]]$SubjectDescriptor,
 
-        [Parameter(Mandatory, ValueFromPipelineByPropertyName)]
+        [Parameter(ValueFromPipelineByPropertyName, ParameterSetName = 'GetMembership')]
         [string]$ContainerDescriptor,
+
+        [Parameter(ValueFromPipelineByPropertyName, ParameterSetName = 'ListMemberships')]
+        [int32]$Depth,
+
+        [Parameter(ValueFromPipelineByPropertyName, ParameterSetName = 'ListMemberships')]
+        [ValidateSet('up', 'down')]
+        [string]$Direction,
 
         [Parameter()]
         [Alias('ApiVersion')]
@@ -66,6 +106,8 @@
         Write-Debug ("CollectionUri: $CollectionUri")
         Write-Debug ("SubjectDescriptor: $($SubjectDescriptor -join ',')")
         Write-Debug ("ContainerDescriptor: $ContainerDescriptor")
+        Write-Debug ("Depth: $Depth")
+        Write-Debug ("Direction: $Direction")
         Write-Debug ("Version: $Version")
 
         Confirm-Default -Defaults ([ordered]@{
@@ -79,18 +121,38 @@
 
     process {
         try {
-            $params = @{
-                Uri     = "$CollectionUri/_apis/graph/memberships/$SubjectDescriptor/$ContainerDescriptor"
-                Version = $Version
-                Method  = 'GET'
+            $queryParameters = [List[string]]::new()
+
+            if ($ContainerDescriptor) {
+                $uri = "$CollectionUri/_apis/graph/memberships/$SubjectDescriptor/$ContainerDescriptor"
+            } else {
+                $uri = "$CollectionUri/_apis/graph/memberships/$SubjectDescriptor"
+
+                if ($Depth) {
+                    $queryParameters.Add("depth=$Depth")
+                }
+                if ($Direction) {
+                    $queryParameters.Add("direction=$Direction")
+                }
             }
 
-            $result = Invoke-AdoRestMethod @params
+            $params = @{
+                Uri             = $uri
+                Version         = $Version
+                QueryParameters = if ($queryParameters.Count -gt 0) { $queryParameters -join '&' } else { $null }
+                Method          = 'GET'
+            }
 
-            [PSCustomObject]@{
-                memberDescriptor    = $result.memberDescriptor
-                containerDescriptor = $result.containerDescriptor
-                collectionUri       = $CollectionUri
+            $results = Invoke-AdoRestMethod @params
+            $memberships = if ($ContainerDescriptor) { @($results) } else { $results.value }
+
+            foreach ($m_ in $memberships) {
+                $obj = [ordered]@{
+                    containerDescriptor = $m_.containerDescriptor
+                    memberDescriptor    = $m_.memberDescriptor
+                    collectionUri       = $CollectionUri
+                }
+                [PSCustomObject]$obj
             }
 
         } catch {