+++ CHANGELOG.md
@@ -19,13 +19,16 @@ All notable changes to this project will be documented in this file.
 
 <br> -->
 
+
 ## [UNRELEASED]
 
 ### Summary
 - _None_
 
 ### What's Changed
+- feat(Get-AdoUser): Add user support (#124)
 - feat(Get-AdoGroup): Add `domain` property to output (#126)
+- fix(Get-AdoUserEntitlement): Correct error handling (#128)
 
 ### Breaking Changes
 - _None_