+++ CHANGELOG.md
@@ -26,8 +26,8 @@ All notable changes to this project will be documented in this file.
 
 ### What's Changed
 - feat(Get-AdoMembership): Add list support (#139)
-- bug(Get-AdoMembership): Set ListMemberships parameter set as default (#144)
 - feat(Get-AdoStorageKey): Resolve a descriptor to a storage key (#142)
+- bug(Get-AdoMembership): Set ListMemberships parameter set as default (#144)
 
 ### Breaking Changes
 - _None_